requirements.txt
openai==1.14.3
python-telegram-bot==20.7
